implement `validate_mask_changes()`, closing #8
function is designed to be used when loading only a subset of an image  - but in principal it can be used for fully-loaded images as well. In that case, it would validate if any mask reaches the edges of the actual image, which might indicate that an image was no shot sufficiently.